Workable version - start
This version works keeping in mind the interests of students and mentors. It takes data from the sheet and uploads to it accordingly as the mentors keep getting assigned. If there isn't anything in the sheet, this code assigns the default value in local machine and if matching is done, updates the "correct" value online.